Resume & JD LLM v3 4/requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
wordcloud==1.9.3